requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
Brotli==1.0.7
CacheControl==0.12.6
cachetools==4.2.0
camelot-py==0.8.2
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
cloudinary==1.22.0
cryptography==3.0
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
distro==1.5.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.8
django-cors-headers==3.4.0
django-cors-middleware==1.5.0
django-heroku==0.3.1
django-htmlmin==0.11.0
django-import-export==2.3.0

djangorestframework==3.11.0
docopt==0.6.2
Elixir==0.7.1
et-xmlfile==1.0.1
firebase-admin==4.5.0
gcloud==0.17.0
ghostscript==0.6
gitignore==0.0.8
google-api-core==1.24.0
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-cloud-core==1.5.0
google-cloud-firestore==2.0.2
google-cloud-storage==1.35.0
google-crc32c==1.1.0
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpcio==1.34.0
gunicorn==20.0.4
html5lib==1.1
httplib2==0.18.1
idna==2.10
install==1.3.3
jdcal==1.4.1
Jinja2==2.11.2
jws==0.1.3
MarkupPy==1.14
MarkupSafe==1.1.1
msgpack==1.0.1
numpy==1.19.1
oauth2client==3.0.0
odfpy==1.4.1
opencv-python==4.4.0.40
opencv-python-headless==4.4.0.46
openpyxl==3.0.4
pandas==1.0.5
pbr==5.4.5

pdfminer.six==20201018
pdfplumber==0.5.25
Pillow==7.2.0
proto-plus==1.13.0
protobuf==3.13.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.4.3
pymongo==3.11.0
PyPDF2==1.26.0
Pyrebase==3.0.27

python-dateutil==2.8.1
python-jwt==2.0.1
pytz==2020.1
PyYAML==5.3.1
requests==2.25.0
requests-toolbelt==0.7.0
rsa==4.6
six==1.15.0
sortedcontainers==2.2.2
soupsieve==2.1
SQLAlchemy==0.7.10
sqlalchemy-migrate==0.13.0
sqlparse==0.3.1
static3==0.7.0
tablib==2.0.0
tabula-py==2.1.1
Tempita==0.5.2
uritemplate==3.0.1
urllib3==1.25.10
Wand==0.6.3
webencodings==0.5.1
whitenoise==5.2.0
xlrd==1.2.0
xlwt==0.7.2
